Change webpages to refer to templates/ for navbar, footer and scripts
This would mean changes to the template will apply to all pages automatically.
To use it, just add the respective include php lines to the top (for navbar) and bottom (footer, script) of the contents of a page.
When the browser loads the page, it will just read the first include and append it into the current document. (try view page source on any page the stuff would be added)